+++ examples/simple/time_series_forecasting/api_forecasting.py
@@ -1,6 +1,5 @@
 import logging
 
-import numpy as np
 import pandas as pd
 
 from fedot.api.main import Fedot
