+++ examples/simple/time_series_forecasting/ts_pipelines.py
@@ -251,15 +251,10 @@ def cgru_pipeline():
 
     Where cgru - convolutional long short-term memory model
     """
-    cgru_params = {'window_size': 29, 'hidden_size': 50, 'learning_rate': 0.004,
-                    'cnn1_kernel_size': 5, 'cnn1_output_size': 32,
-                    'cnn2_kernel_size': 4, 'cnn2_output_size': 32,
-                    'batch_size': 64, 'num_epochs': 3, 'teacher_forcing': 0.8,
-                    'optimizer': 'adam', 'loss': 'mse'}
 
     pip_builder = PipelineBuilder() \
         .add_sequence('lagged', 'ridge', branch_idx=0) \
-        .add_sequence('lagged', ('cgru', cgru_params), branch_idx=1) \
+        .add_sequence('lagged', 'cgru', branch_idx=1) \
         .join_branches('ridge')
 
     pipeline = pip_builder.build()
