+++ test/unit/tasks/test_forecasting.py
@@ -6,13 +6,13 @@ from typing import Optional
 import numpy as np
 import pandas as pd
 import pytest
-from fedot.core.pipelines.pipeline_builder import PipelineBuilder
-from sklearn.metrics import mean_absolute_error, mean_squared_error
 from scipy import stats
+from sklearn.metrics import mean_absolute_error, mean_squared_error
 
-from examples.simple.time_series_forecasting.ts_pipelines import cgru_pipeline
 from fedot.core.data.data import InputData
 from fedot.core.data.data_split import train_test_data_setup
+from fedot.core.operations.evaluation.operation_implementations.models.ts_implementations.arima import \
+    ARIMAImplementation
 from fedot.core.operations.evaluation.operation_implementations.models.ts_implementations.statsmodels import \
     AutoRegImplementation
 from fedot.core.operations.operation_parameters import OperationParameters
@@ -21,10 +21,8 @@ from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.pipelines.ts_wrappers import in_sample_ts_forecast, out_of_sample_ts_forecast
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import Task, TaskTypesEnum, TsForecastingParams
-from fedot.utilities.synth_dataset_generator import generate_synthetic_data
-from fedot.core.operations.evaluation.operation_implementations.models.ts_implementations.arima import \
-    ARIMAImplementation
 from fedot.core.utils import fedot_project_root
+from fedot.utilities.synth_dataset_generator import generate_synthetic_data
 
 np.random.seed(42)
 seed(42)
@@ -291,42 +289,6 @@ def test_multistep_in_sample_forecasting():
     predicted = in_sample_ts_forecast(pipeline=pipeline,
                                       input_data=test_data,
                                       horizon=horizon)
-
-    assert len(predicted) == horizon
-
-
-def test_cgru_forecasting():
-    horizon = 5
-    window_size = 20
-    n_steps = 100
-    train_data, test_data = get_ts_data(n_steps=n_steps + horizon, forecast_length=horizon)
-
-    pipeline = PipelineBuilder().add_node('lagged').add_node('cgru', params = {
-        'input_size': 1,
-        'window_size': window_size,
-        'hidden_size': 100,
-        'learning_rate': 0.0004,
-        'cnn1_kernel_size': 5,
-        'cnn1_output_size': 32,
-        'cnn2_kernel_size': 4,
-        'cnn2_output_size': 32,
-        'batch_size': 64,
-        'num_epochs': 2
-    }).build()
-    pipeline.fit(train_data)
-    predicted = pipeline.predict(test_data).predict[0]
-
-    assert len(predicted) == horizon
-
-
-def test_cgru_in_pipeline():
-    horizon = 5
-    n_steps = 100
-    train_data, test_data = get_ts_data(n_steps=n_steps + horizon, forecast_length=horizon)
-
-    pipeline = cgru_pipeline()
-    pipeline.fit(train_data)
-    predicted = pipeline.predict(test_data).predict[0]
 
     assert len(predicted) == horizon
